add negative tests for login user

+++ src/test/java/site/nomoreparties/stellarburgers/user/LoginUserTests.java
@@ -2,27 +2,47 @@ package site.nomoreparties.stellarburgers.user;
 
 import io.restassured.response.ValidatableResponse;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 public class LoginUserTests {
     private final UserClient client = new UserClient();
     private final UserChecks check = new UserChecks();
+    private User user;
     String accessToken;
+
+    @Before
+    public void setUser() {
+        user = UserGenerator.generateRandomUser();
+        UserClient.createUser(user);
+    }
 
     @Test
     public void loginUserTest() {
-
-        var user = UserGenerator.generateRandomUser();
-        UserClient.createUser(user);
-
         var authdata = AuthData.from(user);
         ValidatableResponse login = client.loginUser(authdata);
         accessToken = check.accertUserLoginSuccessfully(login, user.getEmail(), user.getName());
     }
 
+    @Test
+    public void errorLoginWithWrongEmail() {
+        var authdata = AuthData.wrongEmail(user);
+        ValidatableResponse loginWithWrongEmail = client.loginUser(authdata);
+        check.assertErrorLoginUserWithWrongData(loginWithWrongEmail);
+    }
+
+    @Test
+    public void errorLoginWithWrongPassword() {
+        var authdata = AuthData.wrongPassword(user);
+        ValidatableResponse loginWithWrongPassword = client.loginUser(authdata);
+        check.assertErrorLoginUserWithWrongData(loginWithWrongPassword);
+    }
+
     @After
     public void deleteUser() {
-        ValidatableResponse deleteUser = client.delete(accessToken);
-        check.assertUserDeleteSuccsessfully(deleteUser);
+        if(accessToken != null) {
+            ValidatableResponse deleteUser = client.delete(accessToken);
+            check.assertUserDeleteSuccsessfully(deleteUser);
+        }
     }
 }
